# lib/api-zod/src/generated/types/evaluateMemoryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * QueenSync — Kannaka agent orchestration & resonance control plane
 * OpenAPI spec version: 0.2.0
 */
import type { EvaluateMemoryBodyMetadata } from "./evaluateMemoryBodyMetadata";
import type { EvaluateMemoryBodyType } from "./evaluateMemoryBodyType";

export interface EvaluateMemoryBody {
  type: EvaluateMemoryBodyType;
  content: string;
  /** @nullable */
  agentId?: string | null;
  /** @nullable */
  sourceTaskId?: string | null;
  /** @nullable */
  sourceResonanceId?: string | null;
  metadata?: EvaluateMemoryBodyMetadata;
}